requirements.txt
# Core data processing and modeling libraries
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.5.1
catboost==1.2.3
xgboost==2.1.3
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
python-pptx==0.6.23
streamlit==1.34.0
plotly==5.22.0
